Allow search query string include special characters such as  blank, '?', '&'
Currently, the get URL is not encoded.
If the getURL include blank, it will get nothing.
If the getURL include character '&', oAuth::buildOAuthRawDataKeyValPairs can't split key value pairs correctly.
In my fix, http parameters will be directly sent to a new function oAuth::buildOAuthHttpParameterKeyValPairs then used to be encoded as part of getURL.